bug修复

+++ src/main/java/com/zte/msg/alarmcenter/service/AlarmAbnormalService.java
@@ -4,5 +4,5 @@ import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.zte.msg.alarmcenter.dto.res.AlarmAbnormalResDTO;
 
 public interface AlarmAbnormalService {
-    Page<AlarmAbnormalResDTO> getAlarmAbnormal(String startTime, String endTime, Long systemCode, Long page, Long size);
+    Page<AlarmAbnormalResDTO> getAlarmAbnormal(String startTime, String endTime, Long systemCode, Integer page, Integer size);
 }
